ask-shreyas-rag/backend/requirements.txt
# ── Runtime ────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq==0.9.0
python-dotenv==1.0.1
pydantic==2.7.1
chromadb==0.5.3
httpx==0.27.2

# Embedding model — must match ingest.py (BAAI/bge-small-en-v1.5, 384-dim)
# Removed from runtime in v3 caused total retrieval failure due to dimension mismatch
# with nomic-embed-text (768-dim). BGE is 130MB — fits Render free tier fine.
sentence-transformers==3.0.1
torch==2.3.1
